backend/requirements.txt
# Web framework & server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Supabase client
supabase>=2.4.6

# LlamaIndex ecosystem (RAG orchestration — modular, purpose-built for RAG)
llama-index-core==0.14.16
llama-index-readers-file                  # txt, pdf, csv, json loaders
llama-index-vector-stores-supabase        # pgvector integration
llama-index-embeddings-huggingface        # local HuggingFace embeddings
llama-index-llms-groq                     # Groq/Llama3 LLM integration

# Embeddings — local, no API key needed
sentence-transformers>=3.0.1

# Groq LLM API client (used directly by llama-index-llms-groq)
groq>=0.9.0

# Document loaders / text splitting helpers
unstructured>=0.21.5
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.9.2
